V0.3

+++ internal/db/postgres/users.go
@@ -2,59 +2,75 @@ package postgres
 
 import (
 	"avitoTask/internal/model"
+	"context"
 	"fmt"
+	"github.com/jackc/pgx/v5"
 )
 
-func (c Client) CreateUser(userID model.UserID, segmentsToAdd []model.Slug, segmentsToDelete []model.Slug) error {
-	tx, err := c.db.Begin()
+func (c Client) CreateUser(ctx context.Context, userID model.UserID, segmentsToAdd []model.Slug, segmentsToDelete []model.Slug) error {
+	tx, err := c.db.Begin(ctx)
 	if err != nil {
 		return err
 	}
 	defer func() {
 		if err != nil {
-			tx.Rollback()
+			tx.Rollback(ctx)
 		} else {
-			tx.Commit()
+			tx.Commit(ctx)
 		}
 	}()
 
-	exist, err := c.segmentsExist(segmentsToAdd)
-	if err != nil {
-		return fmt.Errorf("failed to check if segments to add are exist: %w", err)
-	}
-	if !exist {
-		return fmt.Errorf("trying to insert non existing segments: %w", err)
-	}
-
-	q := `
+	if len(segmentsToAdd) != 0 {
+		slugsToAdd := make([]string, len(segmentsToAdd))
+		for i := range segmentsToAdd {
+			slugsToAdd[i] = string(segmentsToAdd[i])
+		}
+		exist, err := c.segmentsExist(ctx, slugsToAdd)
+		if err != nil {
+			return fmt.Errorf("failed to check if segments to add are exist: %w", err)
+		}
+		if !exist {
+			return fmt.Errorf("trying to insert non existing segments: %w", err)
+		}
+		q := `
 		INSERT INTO
 		  users_segments(user_id, slug)
 		VALUES
 		  ($1, $2)
 		ON CONFLICT DO NOTHING
 	`
-	for _, segment := range segmentsToAdd {
-		if _, err := c.db.Exec(q, userID, segment); err != nil {
+		batch := &pgx.Batch{}
+		for _, slugToAdd := range slugsToAdd {
+			batch.Queue(q, userID, slugToAdd)
+		}
+		br := c.db.SendBatch(ctx, batch)
+		ct, err := br.Exec()
+		if err != nil {
 			return fmt.Errorf("failed to insert user segments: %w", err)
 		}
+		defer br.Close()
+		ct.RowsAffected()
 	}
 
-	exist, err = c.segmentsExist(segmentsToDelete)
-	if err != nil {
-		return fmt.Errorf("failed to check if segments exist to delete are: %w", err)
-	}
-	if !exist {
-		return fmt.Errorf("trying to delete non existing segments: %w", err)
-	}
-
-	q = `
+	if len(segmentsToDelete) != 0 {
+		slugsToDelete := make([]string, len(segmentsToDelete))
+		for i := range segmentsToDelete {
+			slugsToDelete[i] = string(segmentsToDelete[i])
+		}
+		exist, err := c.segmentsExist(ctx, slugsToDelete)
+		if err != nil {
+			return fmt.Errorf("failed to check if segments exist to delete are: %w", err)
+		}
+		if !exist {
+			return fmt.Errorf("trying to delete non existing segments: %w", err)
+		}
+		q := `
 		DELETE FROM
 		  users_segments
 		WHERE
-		  user_id = $1 AND slug = $2
+		  user_id = $1 AND slug = any($2)
 	`
-	for _, segment := range segmentsToDelete {
-		if _, err := c.db.Exec(q, userID, segment); err != nil {
+		if _, err := c.db.Exec(ctx, q, userID, slugsToDelete); err != nil {
 			return fmt.Errorf("failed to insert user segments: %w", err)
 		}
 	}
@@ -62,11 +78,7 @@ func (c Client) CreateUser(userID model.UserID, segmentsToAdd []model.Slug, segm
 	return nil
 }
 
-func (c Client) segmentsExist(segments []model.Slug) (bool, error) {
-	slugs := make([]string, len(segments))
-	for i := range segments {
-		slugs[i] = string(segments[i])
-	}
+func (c Client) segmentsExist(ctx context.Context, slugs []string) (bool, error) {
 	q := `
 		SELECT
 		  *
@@ -76,7 +88,7 @@ func (c Client) segmentsExist(segments []model.Slug) (bool, error) {
 		  slug = any($1)
 	`
 
-	rows, err := c.db.Query(q, slugs)
+	rows, err := c.db.Query(ctx, q, slugs)
 	if err != nil {
 		return false, fmt.Errorf("failed to select existing segments: %w", err)
 	}
@@ -86,15 +98,15 @@ func (c Client) segmentsExist(segments []model.Slug) (bool, error) {
 	for rows.Next() {
 		count++
 	}
-
-	if count != len(segments) {
+	fmt.Println(count, len(slugs))
+	if count != len(slugs) {
 		return false, nil
 	}
 
 	return true, nil
 }
 
-func (c Client) SelectActiveSegments(userID model.UserID) ([]model.Slug, error) {
+func (c Client) SelectActiveSegments(ctx context.Context, userID model.UserID) ([]model.Slug, error) {
 	q := `
 		SELECT
 		  slug
@@ -103,7 +115,7 @@ func (c Client) SelectActiveSegments(userID model.UserID) ([]model.Slug, error) 
 		WHERE
 		  user_id = $1
 	`
-	rows, err := c.db.Query(q, userID)
+	rows, err := c.db.Query(ctx, q, userID)
 	if err != nil {
 		return nil, fmt.Errorf("failed to select user active segments: %w", err)
 	}
